Allow generating the private key on the computer
This is to allow creating backups of the private key in case the YubiKey
gets lost or damaged. The word "insecurely" is appended to the name of
the switch and a long warning produced in order to make people who don't
know what they're doing turn away and stop what they're doing.